Python_basics/requirements.txt
appdirs==1.4.4
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==44.0.1
distlib==0.3.9
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.17.0
hello-cli==0.0.2
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
parso==0.8.4
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
prompt_toolkit==3.0.50
ptpython==3.0.29
pycparser==2.22
Pygments==2.19.1
pyproject_hooks==1.2.0
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
SecretStorage==3.3.3
shellingham==1.5.4
tomlkit==0.13.2
trove-classifiers==2025.1.15.22
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
wcwidth==0.2.13
zipp==3.21.0


